requirements.txt
pycryptodome==3.19.0
pywidevine==1.6.0
requests==2.31.0
python-decouple=3.8
